+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   mathfield.
 - **#2805** On mobile, show the math virtual keyboard when using prompts
 - Custom macros are now read when using speakable text to speech.
+- **#2824** Using TypeScript would trigger a compile error about
+  `Cannot find module '../editor/keyboard-layout'`
 
 ## 0.107.0 _2025-08-17_